Move action functions into worker/action.js
